NewReelGen/requirements.txt
openai
python-dotenv
fastapi
uvicorn
jinja2
python-multipart